tests/wpt/web-platform-tests/css/cssom-view/MediaQueryList-addListener-handleEvent.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="flags" content="dom">
<title>CSS Test: CSSOM View MediaQueryList::addListener with handleEvent</title>
<link rel="help" href="https://dom.spec.whatwg.org/#callbackdef-eventlistener">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/matchMedia.js"></script>
<div id="log"></div>
<script>
"use strict";
setup({ allow_uncaught_exception: true });

promise_test(async t => {
    const mql = await createMQL(t);

    let _this;
    let _event;
    const listener = {
        handleEvent(event) {
            _this = this;
            _event = event;
        },
    };

    mql.addListener(listener);
    triggerMQLEvent(mql);
    await waitForChangesReported();

    assert_equals(_this, listener);
    assert_equals(_event.media, mql.media);
    assert_equals(_event.matches, mql.matches);
}, "calls handleEvent method of event listener");

promise_test(async t => {
    const mql = await createMQL(t);

    let calls = 0;
    mql.addListener({
        get handleEvent() {
            calls++;
            return function() {};
        },
    });
    assert_equals(calls, 0);

    triggerMQLEvent(mql);
    await waitForChangesReported();
    assert_equals(calls, 1);

    triggerMQLEvent(mql);
    await waitForChangesReported();
    assert_equals(calls, 2);
}, "looks up handleEvent method on every event dispatch");

promise_test(async t => {
    const mql = await createMQL(t);

    let calls = 0;
    const listener = () => {
        calls++;
    };

    Object.defineProperty(listener, "handleEvent", {
        get: t.unreached_func("handleEvent method should not be looked up on functions"),
    });
    mql.addListener(listener);

    triggerMQLEvent(mql);
    await waitForChangesReported();
    assert_equals(calls, 1);
}, "doesn't look up handleEvent method on callable event listeners");

const uncaught_error_test = async (t, getHandleEvent) => {
    const mql = await createMQL(t);

    let calls = 0;
    mql.addListener({
        get handleEvent() {
            calls++;
            return getHandleEvent();
        },
    });

    const eventWatcher = new EventWatcher(t, window, "error", waitForChangesReported);
    const errorPromise = eventWatcher.wait_for("error");
    triggerMQLEvent(mql);

    const event = await errorPromise;
    assert_equals(calls, 1, "handleEvent property was not looked up");
    throw event.error;
};

promise_test(t => {
    const error = { name: "test" };

    return promise_rejects_exactly(t, error,
        uncaught_error_test(t, () => { throw error; }));
}, "rethrows errors when getting handleEvent");

promise_test(t => {
    return promise_rejects(t, new TypeError(), uncaught_error_test(t, () => false));
}, "throws if handleEvent is falsy and not callable");

promise_test(t => {
    return promise_rejects(t, new TypeError(), uncaught_error_test(t, () => "str"));
}, "throws if handleEvent is thruthy and not callable");
</script>
